feat: Add atomic payment RPCs for concurrent safety, fix UI overflow and profile improvements
- Add atomic PostgreSQL RPCs (initiate_payment_atomic, complete_payment_atomic, fail_payment_atomic) to handle 1000+ concurrent payments safely with FOR UPDATE row locking
- Fix lpad type cast in generate_payment_number (numeric -> integer)
- Add check_fees_locked with 30-min stale lock expiry
- Add database constraints (non-negative balance, unique paynumber)
- Refactor payment_provider to use atomic RPCs instead of multi-step client-side operations
- Fix fee card bottom overflow (4px) in home screen
- Stack Quick Actions buttons vertically in profile screen for mobile
- Hide Switch Account option when only one student exists
- Show 'N/A' instead of 'NULL' for missing student data
- Fix notification time display and profile address formatting

+++ db_sync/setup_payment_safety.sql
@@ -29,7 +29,7 @@ BEGIN
 
   -- Build payment number: extract prefix (e.g., 'FC25/') and pad the number
   v_pay_number := regexp_replace(v_seq.sequid, '\d+$', '') ||
-                  lpad(v_new_no::TEXT, v_seq.seqwidth, '0');
+                  lpad(v_new_no::TEXT, v_seq.seqwidth::INTEGER, '0');
 
   -- Update the sequence atomically
   UPDATE sequence SET seqcurno = v_new_no WHERE seq_id = v_seq.seq_id;
